Move test doubles to a files used for them
