+++ test/test_action.py
@@ -5,25 +5,12 @@ from pyservice import (
     ExpectedKeyNotFoundError,
     PromisedKeyNotFoundError,
 )
+from .test_doubles import add_two
 
 
 @pytest.fixture
 def ctx() -> Context:
     return Context.make()
-
-
-@action()
-def addTwo(ctx: Context) -> Context:
-    n = ctx["n"]
-    ctx["result"] = n + 2
-    return ctx
-
-
-@action(expects=["n"])
-def addThree(ctx: Context) -> Context:
-    n = ctx["n"]
-    ctx["result"] = n + 3
-    return ctx
 
 
 class AddTwo(Action):
@@ -35,23 +22,16 @@ class AddTwo(Action):
 
 def test_fn__can_operate_on_context(ctx: Context) -> None:
     ctx["n"] = 3
-    an_action = addTwo
+    an_action = add_two
     an_action(ctx)
 
     assert ctx["result"] == 5
 
 
-def test_fn__will_fail_without_required_key(ctx: Context) -> None:
-    ctx["x"] = 3
-
-    with pytest.raises(KeyError):
-        addTwo(ctx)
-
-
 def test_fn__will_not_execute_if_in_failed_state(ctx: Context) -> None:
     ctx.fail()
     ctx["n"] = 3
-    addTwo(ctx)
+    add_two(ctx)
 
     assert "result" not in ctx.keys()
