+++ test/test_organizer.py
@@ -1,52 +1,16 @@
 import pytest
-from pyservice import action, Context, Organizer, Organizer2
-from typing import Callable, List
-from .test_doubles import AddTwo, AddTwoWithRollback, AddThree, Fail
-
-
-@action()
-def add_two(ctx: Context) -> Context:
-    n = ctx["n"]
-
-    result = ctx.get("result", n)
-
-    ctx.update(result=result + 2)
-    return ctx
-
-
-@action()
-def fail_context(ctx: Context) -> Context:
-    ctx.fail("Something went wrong...")
-    return ctx
-
-
-@action()
-def skip_rest(ctx: Context) -> Context:
-    ctx.skip("No need to run the rest")
-    return ctx
-
-
-@action()
-def add_three(ctx: Context) -> Context:
-    ctx["result"] += 3
-    return ctx
-
-
-def organizer(ctx: Context) -> Context:
-    actions: List[Callable] = [add_two, add_three]
-    for an_action in actions:
-        an_action(ctx)
-
-    return ctx
-
-
-def organizer2(ctx: Context) -> Context:
-    ctx["result"] = 2
-    actions = [AddTwo, AddThree]
-    for an_action in actions:
-        an_action().execute(ctx)
-
-    return ctx
+from pyservice import Context, Organizer, Organizer2
+from .test_doubles import (
+    add_two,
+    add_three,
+    organizer,
+    fail_context,
+    skip_rest,
+    AddTwo,
+    AddTwoWithRollback,
+    AddThree,
+    Fail,
+)
 
 
 @pytest.fixture
